ドメインロジックの分離
